requirements.txt
torchmetrics>=0.11
tensorboard>=2.11
numpy>=1.23
PyYAML>=6.0
tqdm>=4.63
matplotlib>=3.6